requirements.txt
azure-functions
requests
pillow
